// src/routes/type.ts
export type LeaderBoardDto = 
{
    id:number;
    name: string;
    totalScoreWeekly: number;
    imagePath: string;
}

export type SelfLeaderBoardDto = 
{
    id:number;
    name: string;
    totalScoreWeekly: number;
    imagePath: string;
    userRank:number;
}